Update the completed status in the local storage
